@@ -7,10 +7,17 @@
       "Nebkam\\OdmSearchParam\\": "src"
     }
   },
+  "autoload-dev": {
+    "psr-4": {
+      "Nebkam\\OdmSearchParam\\Tests\\": "tests"
+    }
+  },
   "require": {
-    "php": "^7.0",
-    "doctrine/annotations": "^1.6",
-    "alcaeus/mongo-php-adapter": "^1.1",
-    "doctrine/mongodb-odm": "^1.2"
+    "php": "^7.4",
+    "ext-mongodb": "^1.9",
+    "doctrine/mongodb-odm": "^2.1"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9.5"
   }
 }
