@@ -11,23 +11,25 @@ use Doctrine\Common\Annotations\Annotation\Enum;
 class SearchParam
 	{
 	/**
-	 * @Enum({"string", "string_array", "int", "int_array", "int_gt", "bool", "virtual_bool", "range_int", "range_float", "exists"})
+	 * @Enum({"string", "string_array", "int", "int_array", "int_gt", "bool", "virtual_bool", "range", "range_int", "range_float", "exists"})
 	 */
-	public $type;
+	public ?string $type = null;
 	/**
 	 * @Enum({"from", "to"})
 	 * Used with type `range_int` and `range_float`
 	 */
-	public $direction;
+	public ?string $direction = null;
 	/**
-	 * @var string Explicitly name the field that the property value applies to. Defaults to property name.
+	 * Explicitly name the field that the property value applies to. Defaults to property name.
 	 */
-	public $field;
+	public ?string $field = null;
 	/**
-	 * Method name on the search filter
-	 * that's being called with the query builder and filter property value as an arguments
-	 * and returns the decorated query builder
-	 * @var string
+	 * Method to call with the query builder, filter property value and the whole filter instance as arguments
 	 */
-	public $callback;
+	public ?array $callable = null;
+	/**
+	 * Invert the comparison
+	 * Not all types can be inverted
+	 */
+	public ?bool $invert = null;
 	}
